# openapi.yaml
openapi: 3.1.0
info:
  title: OMNI Pre-Execution Trust & Risk API
  version: 0.2.0
  description: >
    Machine-consumable pre-execution risk evidence for software supply chain and x402 services.
    Paid per request through Circle Gateway nanopayments, including Arc Mainnet.
  contact:
    email: askomni.xyz@gmail.com
  x-guidance: >
    Use OMNI before an agent installs a package, trusts a GitHub repository,
    evaluates a dependency set, or pays an x402 endpoint. Inputs are an exact
    package coordinate, repository owner/name, up to 100 exact package
    coordinates, or a target endpoint URL. Outputs include riskScore,
    scoreStatus, recommendation, evidenceCoverage, bounded scoring-relevant signals and summaries,
    sourceErrors, and deterministic application/json or text/markdown representations. OMNI is
    advisory: caller policy decides whether to proceed, the wallet/runtime
    enforces payment policy, and Circle settles payment. Call OMNI when fresh
    pre-execution evidence is required before software execution or payment.
servers:
  - url: https://api.askomni.xyz
...
paths:
...
  /v1/package/risk:
    get:
      operationId: getPackageRisk
      summary: Assess an exact package version before an agent installs it.
      description: "Price: $0.005 USDC per request. npm assessments combine vulnerability/exploitation evidence with registry supply-chain metadata and optional licensed threat-intelligence matches."
      x-payment-info:
        price:
          mode: fixed
          currency: USDC
          amount: "0.005000"
        protocols:
          - x402: {}
        accepts:
          - scheme: exact
            network: "eip155:5042"
            asset: "0x3600000000000000000000000000000000000000"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:8453"
            asset: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:1"
            asset: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"

      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/Accept' }
        - { in: query, name: ecosystem, required: true, schema: { type: string, minLength: 1, maxLength: 32 }, example: npm }
        - { in: query, name: name, required: true, schema: { type: string, minLength: 1, maxLength: 256 }, example: express }
        - { in: query, name: version, required: true, schema: { type: string, minLength: 1, maxLength: 128 }, example: 5.2.1 }
      responses:
        '200':
          description: Compact machine-readable assessment with explicit source failures and omission counts.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskAssessmentResponse' }
            text/markdown:
              schema: { type: string }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '503': { $ref: '#/components/responses/PaidRequestUnavailable' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
  /v1/repo/risk:
    get:
      operationId: getRepositoryRisk
      summary: Assess GitHub repository security practices.
      description: "Price: $0.01 USDC per request. GitHub repository evidence is bound to the resolved immutable commit; deps.dev provenance is checked against the expected repository/commit where available. Repository evidence from GitHub, OpenSSF Scorecard, dependency resolution, deps.dev provenance, and threat intelligence is attributed independently. Missing evidence affects coverage, scoreStatus, and recommendation gating; it does not itself increase observed repository riskScore."
      x-payment-info:
        price:
          mode: fixed
          currency: USDC
          amount: "0.010000"
        protocols:
          - x402: {}
        accepts:
          - scheme: exact
            network: "eip155:5042"
            asset: "0x3600000000000000000000000000000000000000"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:8453"
            asset: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:1"
            asset: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"

      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/Accept' }
        - { in: query, name: owner, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_.-]{1,100}$' }, example: expressjs }
        - { in: query, name: repo, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_.-]{1,100}$' }, example: express }
      responses:
        '200':
          description: Compact machine-readable assessment with explicit source failures and omission counts.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskAssessmentResponse' }
            text/markdown:
              schema: { type: string }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '503': { $ref: '#/components/responses/PaidRequestUnavailable' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
  /v1/dependencies/risk:
    post:
      operationId: assessDependencies
      summary: Assess up to 100 exact dependency versions.
      description: "Price: $0.05 USDC per request. Packages are evaluated in bounded waves to protect upstream capacity."
      x-payment-info:
        price:
          mode: fixed
          currency: USDC
          amount: "0.050000"
        protocols:
          - x402: {}
        accepts:
          - scheme: exact
            network: "eip155:5042"
            asset: "0x3600000000000000000000000000000000000000"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:8453"
            asset: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:1"
            asset: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"

      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/Accept' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [packages]
              properties:
                packages:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { $ref: '#/components/schemas/PackageCoordinate' }
      responses:
        '200':
          description: Per-package assessments and aggregate summary.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DependencyAssessment' }
            text/markdown:
              schema: { type: string }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '503': { $ref: '#/components/responses/PaidRequestUnavailable' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
  /v1/x402/endpoint/preflight:
    get:
      operationId: preflightX402Endpoint
      summary: Assess an x402 endpoint before an agent decides to call or pay it.
      description: "Price: $0.01 USDC per request. Combines Circle Discovery identity, constrained unpaid handshake evidence, OMNI change history, payout configuration drift, and licensed IOC matches when configured. Listed non-GET endpoints are not invoked."
      x-payment-info:
        price:
          mode: fixed
          currency: USDC
          amount: "0.010000"
        protocols:
          - x402: {}
        accepts:
          - scheme: exact
            network: "eip155:5042"
            asset: "0x3600000000000000000000000000000000000000"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:8453"
            asset: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:1"
            asset: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"

      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/Accept' }
        - { in: query, name: url, required: true, schema: { type: string, format: uri, maxLength: 2048 } }
      responses:
        '200':
          description: "Endpoint risk assessment with source-attributed evidence and explicit source failures. preflightContext is observational/advisory: it reports the payment configuration observed during preflight and is not an authorization object."
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/X402EndpointPreflight' }
            text/markdown:
              schema: { type: string }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '503': { $ref: '#/components/responses/PaidRequestUnavailable' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
  /v1/agent/risk:
    get:
      operationId: getAgentRisk
      summary: Assess an ERC-8004 agent's identity, reputation, and service evidence before invoking or paying it.
      description: >
        Price: $0.05 USDC per request.
        Fetches identity from ERC-8004 IdentityRegistry (Ethereum Mainnet + Base Mainnet),
        scans ReputationRegistry for bounded on-chain feedback (NewFeedback / FeedbackRevoked events),
        and fetches the agent card from the registration URI for service evidence.
        A missing or unregistered agent is not itself a malicious signal: it returns
        scoreStatus=insufficient_evidence and recommendation=manual_review.
        The optional targetUrl parameter is an opt-in caller-supplied URL check for redirect-to-private
        detection; it is only probed if it matches a service endpoint advertised in the verified agent card.
      x-payment-info:
        price:
          mode: fixed
          currency: USDC
          amount: "0.050000"
        protocols:
          - x402: {}
        accepts:
          - scheme: exact
            network: "eip155:5042"
            asset: "0x3600000000000000000000000000000000000000"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:8453"
            asset: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
          - scheme: exact
            network: "eip155:1"
            asset: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48"
            payTo: "0xd5154d79b52a5980e7b0e806f5e4bf3dca3798b5"
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/Accept' }
        - in: query
          name: agentId
          required: true
          description: ERC-8004 agent token id. Accepts decimal or 0x-prefixed hex string.
          schema:
            type: string
            pattern: '^(0x[0-9a-fA-F]{1,64}|[0-9]{1,78})$'
          example: "42"
        - in: query
          name: targetUrl
          required: false
          description: >
            Optional HTTPS URL to probe for redirect-to-private. Only probed when it matches
            a service endpoint advertised in the verified agent card. Caller-supplied URLs that
            are not advertised in the agent card are rejected before any probe (no risk increase).
          schema:
            type: string
            format: uri
            maxLength: 2048
      responses:
        '200':
          description: Agent risk assessment with ERC-8004 identity, reputation, and service evidence.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRiskResponse' }
            text/markdown:
              schema: { type: string }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '503': { $ref: '#/components/responses/PaidRequestUnavailable' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
  /.well-known/x402:
    get:
      operationId: getX402DiscoveryManifest
      summary: Free public x402 discovery manifest for agent discoverability.
      description: |
        Returns a machine-readable x402 discovery manifest describing supported payment
        networks and paid resources. This endpoint is FREE (no payment required).

        The networks listed are the currently supported Circle Gateway payment kinds,
        discovered live from the configured facilitator. Payment terms (networks,
        amounts, assets) can change. The actual PAYMENT-REQUIRED challenge from a paid
        request is authoritative at execution time.
      responses:
        '200':
          description: x402 discovery manifest.
          content:
            application/json:
              schema:
                type: object
                required: [x402Version, service, openapi, paymentTerms, networks, resources]
                properties:
                  x402Version: { type: integer, enum: [2] }
                  service: { type: string }
                  description: { type: string }
                  openapi: { type: string, format: uri }
                  paymentTerms: { type: string, enum: [live_402_authoritative] }
                  networks:
                    type: array
                    items: { type: string }
                  resources:
                    type: array
                    items:
                      type: object
                      required: [method, resource, price]
                      properties:
                        method: { type: string }
                        resource: { type: string, format: uri }
                        price:
                          type: object
                          required: [currency, amount]
                          properties:
                            currency: { type: string }
                            amount: { type: string }
        '503':
          description: Gateway discovery is unavailable and no cached capability exists.
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
                    enum: [discovery_unavailable, gateway_discovery_unavailable]
components:
  headers:
    PaymentResponse:
      description: Optional base64-encoded JSON Circle x402 payment receipt emitted for the original successful settlement. OMNI durable replays intentionally omit this header rather than fabricating a new payment receipt. The decoded JSON contains success (boolean), transaction (string), network (string), and payer (string).
      schema:
        type: string
  parameters:
    Accept:
      name: Accept
      in: header
      required: false
      description: Response representation. Use application/json for compact machine/agent decisions or text/markdown for a concise human summary. Choose one representation per request; errors remain JSON. Replaying a completed Idempotency-Key may switch representation without another assessment or payment.
      schema:
        type: string
        default: application/json
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional for an initial unpaid x402 discovery/challenge request. Required when submitting PAYMENT-SIGNATURE and for durable paid execution/replay semantics. It must be a UUID v4 and must be reused unchanged for retries. Reusing it for different validated input returns a conflict.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89a-fA-F][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
  responses:
    InvalidRequest:
      description: Request failed validation before payment was accepted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    PaymentRequired:
      description: Payment is required. Read the PAYMENT-REQUIRED response header and retry with a valid x402 payment signature.
      headers:
        PAYMENT-REQUIRED:
          required: true
          schema: { type: string }
    IdempotencyConflict:
      description: The Idempotency-Key is already bound to different validated route input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotAcceptable:
      description: The requested response representation is not supported or has zero acceptable quality.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    PaidRequestUnavailable:
      description: Replica admission capacity is exhausted, or durable paid-request state/recovery is unavailable. The server fails closed and does not create another payment.
      headers:
        Retry-After:
          schema: { type: string, example: '1' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    PackageCoordinate:
      type: object
      additionalProperties: false
      required: [ecosystem, name, version]
      properties:
        ecosystem: { type: string, minLength: 1, maxLength: 32 }
        name: { type: string, minLength: 1, maxLength: 256 }
        version: { type: string, minLength: 1, maxLength: 128 }
    Subject:
      type: object
      required: [type, id]
      properties:
        type: { type: string, enum: [package, repository, dependency_set, x402_endpoint, agent] }
        id: { type: string }
...
    RiskDimensions:
      type: object
      required: [knownVulnerabilities, knownExploitation, packageSupplyChain, repositorySecurityPractices, maliciousInfrastructure, serviceIdentity, paymentConfigurationRisk, endpointOperationalRisk]
      properties:
        knownVulnerabilities: { $ref: '#/components/schemas/RiskDimensionLevel' }
        knownExploitation: { $ref: '#/components/schemas/RiskDimensionLevel' }
        packageSupplyChain: { $ref: '#/components/schemas/RiskDimensionLevel' }
        repositorySecurityPractices: { $ref: '#/components/schemas/RiskDimensionLevel' }
        maliciousInfrastructure: { $ref: '#/components/schemas/RiskDimensionLevel' }
        serviceIdentity: { $ref: '#/components/schemas/RiskDimensionLevel' }
        paymentConfigurationRisk: { $ref: '#/components/schemas/RiskDimensionLevel' }
        endpointOperationalRisk: { $ref: '#/components/schemas/RiskDimensionLevel' }
...
    EvidenceCoverage:
      type: object
      additionalProperties: false
      required: [modelVersion, resolvedWeight, applicableWeight, sources]
      properties:
        modelVersion: { type: string }
        resolvedWeight: { type: number, minimum: 0 }
        applicableWeight: { type: number, minimum: 0 }
        sources:
          type: array
          items: { $ref: '#/components/schemas/EvidenceCoverageSource' }
...
    ResponseOmissions:
      type: object
      additionalProperties: false
      required: [evidenceDetailsOmitted, dependencyDetailsOmitted]
      description: Public representation omission counts. For x402 preflight, paymentOptionsOmitted reports options beyond the maximum of 8 retained public options; a nonzero count makes an unmatched selected option insufficient context rather than a definitive mismatch.
      properties:
        evidenceDetailsOmitted: { type: integer, minimum: 0 }
        dependencyDetailsOmitted: { type: integer, minimum: 0 }
        signalsOmitted: { type: integer, minimum: 0 }
        sourceErrorsOmitted: { type: integer, minimum: 0 }
        packagesOmitted: { type: integer, minimum: 0 }
        coverageSourcesOmitted: { type: integer, minimum: 0 }
        paymentOptionsOmitted: { type: integer, minimum: 0 }
    CompactRiskAssessmentFields:
      type: object
      required: [subject, policyVersion, scoreStatus, recommendation, riskScore, evidenceCoverage, dimensions, signals, sourceErrors, omissions, assessedAt, freshness]
      properties:
        subject: { $ref: '#/components/schemas/Subject' }
        policyVersion: { type: string }
        scoreStatus: { type: string, enum: [measured, measured_partial, insufficient_evidence] }
        recommendation: { type: string, enum: [proceed, proceed_with_caution, manual_review, do_not_proceed] }
        riskScore: { type: integer, minimum: 0, maximum: 100 }
        evidenceCoverage: { type: number, minimum: 0, maximum: 1 }
        coverage: { $ref: '#/components/schemas/EvidenceCoverage' }
        dimensions: { $ref: '#/components/schemas/RiskDimensions' }
        signals:
          type: array
          maxItems: 32
          items: { $ref: '#/components/schemas/CompactSignal' }
        repositorySummary: { $ref: '#/components/schemas/RepositoryRiskSummary' }
        maliciousPackageObservations:
          type: object
          additionalProperties: false
          required: [observed, ids]
          properties:
            observed: { type: integer, minimum: 0 }
            ids:
              type: array
              maxItems: 16
              items: { type: string }
            idsOmitted: { type: integer, minimum: 0 }
        sourceErrors:
          type: array
          maxItems: 16
          items: { type: string }
        omissions: { $ref: '#/components/schemas/ResponseOmissions' }
        assessedAt: { type: string, format: date-time }
        freshness:
          type: object
          additionalProperties: false
          required: [oldestEvidenceAt, newestEvidenceAt]
          properties:
            oldestEvidenceAt: { type: [string, 'null'], format: date-time }
            newestEvidenceAt: { type: [string, 'null'], format: date-time }
            expiresAt: { type: string, format: date-time }
    RiskAssessmentResponse:
      allOf:
        - $ref: '#/components/schemas/CompactRiskAssessmentFields'
      unevaluatedProperties: false
      description: Compact machine-readable JSON response. Raw evidence, provider payloads, Markdown, and full dependency details are intentionally omitted; omission counts are explicit.
...
    X402PreflightContext:
      type: object
      additionalProperties: false
      required: [resource, paymentOptions]
      description: Public observational/advisory record of the preflight payment configuration. Up to 8 observed options are retained; additional options are reported by the top-level omissions.paymentOptionsOmitted field. It is not an authorization object and does not bind any settlement.
      properties:
        resource: { type: string, format: uri }
        paymentOptions:
          type: array
          maxItems: 8
          items: { $ref: '#/components/schemas/ObservedPaymentRequirement' }
    X402EndpointPreflight:
      type: object
      allOf:
        - $ref: '#/components/schemas/CompactRiskAssessmentFields'
        - type: object
          required: [preflightContext]
          properties:
            preflightContext: { $ref: '#/components/schemas/X402PreflightContext' }
      unevaluatedProperties: false
      description: Top-level successful JSON x402 preflight response with compact risk fields and bounded observed preflight context.
    DependencyAssessment:
      type: object
      required: [packages, summary, assessedAt, omissions]
      properties:
        packages:
          type: array
          maxItems: 16
          items: { $ref: '#/components/schemas/RiskAssessmentResponse' }
        summary:
          type: object
          required: [count, worstRiskScore, recommendations]
          properties:
            count: { type: integer, minimum: 1, maximum: 100 }
            worstRiskScore: { type: integer, minimum: 0, maximum: 100 }
            recommendations:
              type: object
              additionalProperties: { type: integer, minimum: 0 }
        assessedAt: { type: string, format: date-time }
        omissions: { $ref: '#/components/schemas/ResponseOmissions' }
...
    AgentRisk:
      type: object
      additionalProperties: false
      required: [agentId, primaryChainId, dimensions, chainEvidence]
      description: >
        Agent-specific risk extension. Present only when subject.type === "agent".
        agentId is the token id as a decimal string.
        registered=false is not a malicious signal: it produces scoreStatus=insufficient_evidence.
      properties:
        agentId: { type: string, description: "ERC-8004 token id, decimal string." }
        primaryChainId: { type: integer, description: "Chain where the agent was first found registered." }
        agentWallet: { type: string }
        registrationUri: { type: string, format: uri }
        agentName: { type: string }
        agentDescription: { type: string }
        dimensions: { $ref: '#/components/schemas/AgentRiskDimensions' }
        chainEvidence:
          type: array
          items: { $ref: '#/components/schemas/AgentChainIdentityResult' }
        reputationSummary: { $ref: '#/components/schemas/AgentReputationSummary' }
        services:
          type: array
          items: { $ref: '#/components/schemas/AgentServiceObservation' }
        targetUrlVerified:
          type: boolean
          description: True when the caller-supplied targetUrl matches a service endpoint in the verified agent card.
        targetUrlRedirectsToPrivate:
          type: boolean
          description: True when OMNI probed targetUrl and observed a redirect to a private IP. Only set when targetUrl was advertised in the verified agent card.
    AgentRiskResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/CompactRiskAssessmentFields'
        - type: object
          required: [agentRisk]
          properties:
            agentRisk: { $ref: '#/components/schemas/AgentRisk' }
      unevaluatedProperties: false
      description: >
        Compact machine-readable JSON response for an ERC-8004 agent risk assessment.
        The standard CompactRiskAssessmentFields carry the base risk score and coverage;
        the agentRisk extension carries agent-specific identity, reputation, and service evidence.
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        retryable: { type: boolean }